tscsmethods

tscsmethods.att

tscsmethods.caliper

tscsmethods.caliper!

tscsmethods.checkmatchnums

tscsmethods.countmemb

tscsmethods.countmemb

tscsmethods.estsum

tscsmethods.getbalance_restricted

tscsmethods.getvarind

tscsmethods.invert_dict

tscsmethods.make_check_sample!

tscsmethods.makemats!

tscsmethods.matchdistances

tscsmethods.missingmats

tscsmethods.outputprocess

tscsmethods.plot_att

tscsmethods.plot_att

tscsmethods.plot_att

tscsmethods.plot_balance

tscsmethods.plot_mdistances

tscsmethods.refine!

tscsmethods.weeklyattci

tscsmethods.witcal

tscsmethods.att — Method

att(om, wm, trtnums)

this will need to be adapted for missingness, and we will need a new way to get the treated unit number in the presence of missingness, which will vary across f

tscsmethods.caliper! — Method

tscsmethods.caliper — Method

caliper(cdict::Dict, matches::DataFrame)

Enforce a caliper on desired set of covariates, or overall Mahalanabois distance.

Examples ≡≡≡≡≡≡≡≡≡≡

cdict = Dict(:cumdeathrte => 0.5, :first_case => 0.5); matchescal = caliper(cdict, matches);

tscsmethods.checkmatchnums — Method

checkmatchnums(mm)

get the number of matches for each treated observation

tscsmethods.countmemb — Method

faster version when length is known

tscsmethods.countmemb — Method

faster than countmap()

tscsmethods.estsum — Method

helper to sum the bootstrap distributions over some f range

tscsmethods.getbalance_restricted — Method

standardizing standard deviation: this is calculated from the full dataset still, not the stratum subset

tscsmethods.getvarind — Method

getvarind()

get the indices of a unit over time time periode

tscsmethods.invert_dict — Function

https://stackoverflow.com/questions/40694157/how-to-reverse-a-dictionary-in-julia

tscsmethods.make_check_sample! — Method

make_check_sample!(units, clusters, treatedunits, x)

ensure that there are treated units in the resample

tscsmethods.makemats! — Method

this is the new, faster function to make the outcome and dwit mats we cannot simply add the wits to the dataframe, since we are keeping them disaggregated

tscsmethods.matchdistances — Method

will operate on (i - 1) * length(unid) + 1 : length(unid) * i

tscsmethods.missingmats — Method

missingmats(om, wm)

output:

adjusted om

adjusted wm

bl: black list of treated observations with treatment period (f)

tscsmethods.outputprocess — Method

outputprocess(bootests, ests, Fset)

process the bootstrap and estimation output into a dataframe for easy access and plotting

tscsmethods.plot_att — Method

plot_att(atts, estimator; savename = "")

tscsmethods.plot_att — Method

plot_att(
  atts,
  stratvar::Symbol,
  savename = "",
  xinch = 15inch,
  yinch = 6inch,
  treatment = :treatment
)

Plot restricted estimates, without key for strata.

tscsmethods.plot_att — Method

plot_att(
  atts,
  stratvar::Symbol,
  stratdict::Dict;
  savename = "",
  xinch = 15inch,
  yinch = 6inch,
  treatment = :treatment
)

Plot restricted estimates, with key for strata.

tscsmethods.plot_balance — Method

stratdict::Dict points from stratvar to desired label

tscsmethods.plot_mdistances — Method

plot_mdistances(matches5, calvars; savenme = "")

histogram plot the maha distances: overall, and for each identified caliper variable

if savenme is left empty, plot will display but not save to file

tscsmethods.refine! — Method

for separate refinement on a given match set this avoids time-consuming recalculation of the possible match set

tscsmethods.weeklyattci — Method

gives separate sum for each week, respecting days of week, for f window note that, depending on f window, weeks may have different numbers of days

tscsmethods.witcal — Method

witcal(mu, tu, j, mc)

get the disaggregated weight value for an observation. this function is only applied to observations that are non-zero. (although else statement implies that it could.)

Arguments ≡≡≡≡≡≡≡≡≡≡≡

mu: match unit

tu: treated unit for mu

j: vector position for relative timepoint, j == 1 => tpoint, j > 1 => f in F

mc: number of matches to the tu (of which mu is one)